+++ Cesium3DTilesSelection/src/ImplicitOctreeLoader.h
@@ -53,11 +53,6 @@ public:
       Cesium3DTilesContent::SubtreeAvailability&& subtreeAvailability);
 
 private:
-  static std::string resolveUrl(
-      const std::string& baseUrl,
-      const std::string& urlTemplate,
-      const CesiumGeometry::OctreeTileID& octreeID);
-
   std::string _baseUrl;
   std::string _contentUrlTemplate;
   std::string _subtreeUrlTemplate;